feat: 리뷰 반영
- package 분리
- car 추출 로직 inline 방식으로 변경
- 불필요한 class 내 companion object 삭제 처리
