+++ src/main/kotlin/calculator/Person.kt
@@ -1,3 +1,3 @@
-package study
+package calculator
 
 data class Person(val name: String, val age: Int, var nickname: String? = name)
