+++ src/main/kotlin/racing/CarRacingSimulator.kt
@@ -1,11 +1,15 @@
-import Constants.RESULT_MESSAGE
-import RacingGame.getWinner
+package racing
+
+import racing.Constants.RESULT_MESSAGE
+import racing.gamerule.RacingGame
+import racing.gamerule.RacingGame.getWinner
+import racing.gamerule.RandomMovingStrategy
 
 fun main() {
     val inputParams = CarRacingInputHandler.input()
     CarRacingInputHandler.validate(inputParams.numberOfCars, inputParams.numberOfTries, inputParams.carNames)
 
-    val cars = CarRacingInputHandler.getCars(inputParams.carNames)
+    val cars = CarRacingInputHandler.getCars(inputParams)
 
     println(RESULT_MESSAGE)
