+++ src/main/kotlin/racing/InputParams.kt
@@ -1,1 +1,3 @@
+package racing
+
 class InputParams(val numberOfCars: String, val numberOfTries: String, val carNames: List<String>)
