+++ src/main/kotlin/racing/MovingStatus.kt
@@ -1,3 +1,5 @@
+package racing
+
 enum class MovingStatus {
     MOVE,
     STOP
