+++ src/main/kotlin/racing/gamerule/MovingStrategy.kt
@@ -1,3 +1,5 @@
+package racing.gamerule
+
 fun interface MovingStrategy {
     fun isMovable(): Boolean
 }
