+++ src/main/kotlin/racing/gamerule/RacingGame.kt
@@ -1,3 +1,7 @@
+package racing.gamerule
+
+import racing.Car
+
 object RacingGame {
     fun play(cars: List<Car>, movingStrategy: MovingStrategy) {
         cars.forEach { car ->
@@ -11,6 +15,6 @@ object RacingGame {
 
     fun getWinner(cars: List<Car>): List<String> {
         val max = cars.maxOf { it.getMoved() }
-        return cars.filter { it.getMoved().equals(max) }.map { it.carName }
+        return cars.filter { it.getMoved() == max }.map { it.carName }
     }
 }
