+++ src/test/kotlin/calculator/CalculatorTest.kt
@@ -1,10 +1,9 @@
-
+package calculator
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatExceptionOfType
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.ValueSource
-import study.Calculator
 
 class CalculatorTest {
