+++ src/test/kotlin/calculator/PersonKoTest.kt
@@ -1,7 +1,8 @@
+package calculator
+
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.inspectors.forAll
 import io.kotest.matchers.shouldBe
-import study.Person
 
 class PersonKoTest : StringSpec({
