+++ src/test/kotlin/calculator/PersonTest.kt
@@ -1,6 +1,7 @@
+package calculator
+
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
-import study.Person
 
 class PersonTest {
