+++ src/test/kotlin/calculator/StringKoTest.kt
@@ -1,3 +1,5 @@
+package calculator
+
 import io.kotest.core.spec.style.AnnotationSpec
 import io.kotest.matchers.shouldBe
