+++ src/test/kotlin/calculator/StringTest.kt
@@ -1,4 +1,4 @@
-package study
+package calculator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
